Cleaned up generated index.js and package.js files
The generated package.js and index.js files for the npm-container package look
much cleaner now. The ending comment with the line number is removed and
am using js-beautify to make it look pretty.

Clarified:
* what adding the package.json to the npm-container package is fore

Removed:
* deleting packages/npm directory. I imagine this was providing some sort of
backwards compatibility, but when woring on the package locally it would delete
the directory right from under you while you're working on it. No longer!