Removed process.exit call
There's no need for the build process to stop after making the npm-container
directory. Meteor will automatically restart the build when it detects the
addition to .meteor/packages. I've tested that the declared npm modules in
package.json are brought in when running `meteor` in development and `meteor
build` for production